added create and view sale orders

+++ customer/views.py
@@ -1,8 +1,9 @@
 import json
-from django.shortcuts import render, redirect, HttpResponse
+from django.http import Http404
+from django.shortcuts import get_object_or_404, render, redirect, HttpResponse
 from .models import Customer
 from django.core.paginator import Paginator
-from django.db.models import Q  # New
+from django.db.models import Q 
 from django.db.models.functions import Lower
 
 
@@ -11,10 +12,10 @@ def customer(request):
     # For searching the Customers
     search_customer = request.GET.get('search_customer')
     if search_customer:
-        customer = Customer.objects.filter(Q(name__icontains=search_customer))
+        customer = Customer.objects.exclude(id=1).filter(Q(name__icontains=search_customer))
     else:
         # If not searched, return default posts
-        customer = Customer.objects.all().order_by(Lower("name"))
+        customer = Customer.objects.all().exclude(id=1).order_by(Lower("name"))
     
     p = Paginator(customer, 12)  # creating a paginator object
     # getting the desired page number from url
@@ -34,12 +35,11 @@ def add_customer(request):
     if request.method == "POST":
         instance = Customer()
         instance.name = request.POST['name']
-        if (request.POST['gstin'] != ''):
-            instance.gstin = request.POST['gstin']
+        instance.gstin = request.POST['gstin']
+        instance.address1 = request.POST['address1']
+        instance.address2 = request.POST['address2']
         if (request.POST['mob_no'] != ''):
             instance.mob_no = request.POST['mob_no']
-        if (request.POST['address'] != ''):
-            instance.address = request.POST['address']
         if (request.POST['email'] != ''): 
             instance.email = request.POST['email']
         if bool(request.FILES.get('pic', False)) == True:
@@ -49,10 +49,14 @@ def add_customer(request):
     return render(request, 'customer/addCustomer.html', {'sb':4})
 
 def edit_customer(request, customer_id):
-    customer_instance = Customer.objects.get(pk=customer_id)
+    if customer_id == 1:
+        raise Http404
+    customer_instance = get_object_or_404(Customer, pk=customer_id)
+    
     if request.method == 'POST':
         
-        customer_instance.name = request.POST['name']
+        if (request.POST['name'] != ''):
+            customer_instance.name = request.POST['name']
 
         if (request.POST['gstin'] != ''):
             customer_instance.gstin = request.POST['gstin']
@@ -60,8 +64,11 @@ def edit_customer(request, customer_id):
         if (request.POST['mob_no'] != ''):
             customer_instance.mob_no = request.POST['mob_no']
             
-        if (request.POST['address'] != ''):
-            customer_instance.address = request.POST['address']
+        if (request.POST['address1'] != ''):
+            customer_instance.address1 = request.POST['address1']
+            
+        if (request.POST['address2'] != ''):
+            customer_instance.address2 = request.POST['address2']
             
         if (request.POST['email'] != ''):
             customer_instance.email = request.POST['email']
@@ -74,12 +81,13 @@ def edit_customer(request, customer_id):
     return render(request, 'customer/editCustomer.html', {'sb':4, 'customer':customer_instance})
 
 def del_customer(request, customer_id):
-    customer_instance = Customer.objects.get(pk=customer_id)
+    if customer_id == 1:
+        raise Http404
+    customer_instance = get_object_or_404(Customer, pk=customer_id)
     customer_instance.delete()
     return redirect('customer:customer')
 
 def get_customer(request, customer_id):
-    
     customer = Customer.objects.get(pk=customer_id)
     customer_dict = {}
     customer_dict['name'] = customer.name
@@ -87,6 +95,7 @@ def get_customer(request, customer_id):
     customer_dict['gstin'] = customer.gstin
     customer_dict['mob_no'] = customer.mob_no
     customer_dict['email'] = customer.email
-    customer_dict['address'] = customer.address
+    customer_dict['address1'] = customer.address1
+    customer_dict['address2'] = customer.address2
     
     return HttpResponse(json.dumps(customer_dict), content_type="application/json")
